fix: 完善MySQL 8.0 collation的支持(#500)

+++ store/mockstore/mocktikv/cop_handler_dag.go
@@ -697,6 +697,6 @@ func fieldTypeFromPBColumn(col *tipb.ColumnInfo) *types.FieldType {
 		Flen:    int(col.GetColumnLen()),
 		Decimal: int(col.GetDecimal()),
 		Elems:   col.Elems,
-		Collate: mysql.Collations[uint8(col.GetCollation())],
+		Collate: mysql.Collations[uint16(col.GetCollation())],
 	}
 }
